+++ radio/src/telemetry/telemetry.h
@@ -38,7 +38,10 @@
   #include "flysky_ibus.h"
 #endif
 
-extern uint8_t telemetryStreaming; // >0 (true) == data is streaming in. 0 = no data detected for some time
+//OW
+//extern uint8_t telemetryStreaming; // >0 (true) == data is streaming in. 0 = no data detected for some time
+extern uint16_t telemetryStreaming; // >0 (true) == data is streaming in. 0 = no data detected for some time
+//OWEND
 
 inline bool TELEMETRY_STREAMING()
 {
